requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
bboxtools>=1.0.4,<2
cmake==3.31.6
einops>=0.8.1,<0.9
hydra-core>=1.3.2,<2
imageio>=2.37.0,<3
fvcore
pillow
matplotlib>=3.10.3,<4
ninja
opencv-python>=4.11.0.86,<5
rich>=14.0.0,<15
scipy>=1.15.3,<2
scikit-image>=0.25.2,<0.26
triton==3.2.0
wandb>=0.19.11,<0.20
torch 
torchvision 
torchaudio
iopath>=0.1.10
pybind11==2.13.6
